pyproject.toml
[tool.poetry]
name = "fca-init"
version = "0.1.0"
description = ""
authors = ["canarypwn"]

[tool.poetry.dependencies]
python = "^3.10"
dtaidistance = "^2.3.11"
aiohttp = "^3.9.3"
tensorflow = "^2.16.1"
matplotlib = "^3.8.3"
pandas = "^2.2.1"
scipy = "^1.12.0"
scikit-learn = "^1.4.1"
statsmodels = "^0.14.1"
dtwalign = "^0.1.1"
fastdtw = "^0.3.4"
sklearn-som = "^1.1.0"
kneed = "^0.8.5"
openpyxl = "^3.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


